backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib
bcrypt
sqlalchemy
psycopg2-binary
redis
openai
langchain
langgraph
pypdf2
python-dotenv
pydantic
httpx
unstructured
llama-index 